Create Fibonacci Series
This code can be used for displaying fibonacci series. Fibonacci series is a series in which first to digits are 0 & 1 given by us and the coming digits are sum of last two digits. output is shown as follows:-0 1 1 2 3 5 8....and so on till n terms ;where n is to be decided by user.