Create Demonstration of gets() and puts() in c
This code can be used to demonstrate use of gets and puts in c language. The gets() function in C++ reads characters from stdin and stores them until a newline character is found or end of file occurs(reading input). The puts() function in C/C++ is used to write a line or string to the output stream(displaying input).